abstract priority avatar entities, optimize avatar animation updates
